+++ src/BasicRegisterInterface.php
@@ -6,6 +6,7 @@ namespace DigipolisGent\Flanders\BasicRegisters;
 
 use DigipolisGent\Flanders\BasicRegisters\Service\AddressServiceInterface;
 use DigipolisGent\Flanders\BasicRegisters\Service\MunicipalityNameServiceInterface;
+use DigipolisGent\Flanders\BasicRegisters\Service\StreetNameServiceInterface;
 
 /**
  * Container of all BasicRegister services.
@@ -25,4 +26,11 @@ interface BasicRegisterInterface
      * @return \DigipolisGent\Flanders\BasicRegisters\Service\MunicipalityNameServiceInterface
      */
     public function municipalityName(): MunicipalityNameServiceInterface;
+
+    /**
+     * Get the street name related service.
+     *
+     * @return \DigipolisGent\Flanders\BasicRegisters\Service\StreetNameServiceInterface
+     */
+    public function streetName(): StreetNameServiceInterface;
 }