+++ src/Uri/StreetNameListUri.php
@@ -1,0 +1,19 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DigipolisGent\Flanders\BasicRegisters\Uri;
+
+/**
+ * Uri where the street names (streetnamen) method is located.
+ */
+class StreetNameListUri extends AbstractUriWithQuery
+{
+    /**
+     * @inheritDoc
+     */
+    protected function getPath(): string
+    {
+        return 'straatnamen';
+    }
+}