+++ tests/Response/StreetNameListResponseTest.php
@@ -1,0 +1,28 @@
+<?php
+
+declare(strict_types=1);
+
+namespace DigipolisGent\Tests\Flanders\BasicRegisters\Response;
+
+use DigipolisGent\Flanders\BasicRegisters\Response\StreetNameListResponse;
+use DigipolisGent\Flanders\BasicRegisters\Value\Street\StreetNames;
+use PHPUnit\Framework\TestCase;
+
+/**
+ * @covers \DigipolisGent\Flanders\BasicRegisters\Response\StreetNameListResponse
+ */
+class StreetNameListResponseTest extends TestCase
+{
+    /**
+     * Response is created with Addresses collection.
+     *
+     * @test
+     */
+    public function responseHasStreetNamesCollection(): void
+    {
+        $streetNames = new StreetNames();
+        $response = new StreetNameListResponse($streetNames);
+
+        $this->assertSame($streetNames, $response->streetNames());
+    }
+}