+++ CHANGELOG.md
@@ -39,5 +39,7 @@ All Notable changes to `digipolisgent/flanders-basicregisters` package.
 * Added service method to get (partial) addresses that match filter(s).
 * Added service method to get a list of municipality names.
 * Added service method to get the details of a single municipality name.
+* Added service method to get a list of street names.
+* Added service method to get the details of a single street name.
 
 [Unreleased]: https://github.com/digipolisgent/php_package_dg-flanders-basicregisters/compare/master...develop